wijn_fitter/requirements.txt
pandas==1.1.1
sklearn==0.0
pygithub==1.55
